error handling in user controller

+++ src/user/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './schemas/user.schema';
@@ -9,6 +13,12 @@ import * as bcrypt from 'bcrypt';
 export class UserService {
   constructor(@InjectModel(User.name) private userModel: Model<User>) {}
   async create(createUserDto: CreateUserDto): Promise<User> {
+    const isEmailExist = await this.userModel
+      .findOne({ email: createUserDto.email })
+      .exec();
+    if (isEmailExist) {
+      throw new ConflictException('Email already exists');
+    }
     const hashedPassword = await bcrypt.hash(createUserDto.password, 10);
     const createdUser = new this.userModel({
       ...createUserDto,
@@ -22,10 +32,22 @@ export class UserService {
   }
 
   async findOne(id: string) {
-    return await this.userModel.findById(id).exec();
+    const user = await this.userModel.findById(id).exec();
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    return user;
+  }
+
+  async findOneByEmail(email: string) {
+    return await this.userModel.findOne({ email }).exec(); // findOne
   }
 
   async update(id: string, updateUserDto: UpdateUserDto): Promise<User | null> {
+    const user = await this.userModel.findById(id).exec();
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
     return await this.userModel
       .findByIdAndUpdate(id, updateUserDto, {
         new: true,
@@ -34,11 +56,11 @@ export class UserService {
   }
 
   async remove(id: string): Promise<{ deleted: boolean; message?: string }> {
-    try {
-      await this.userModel.findByIdAndDelete(id).exec();
-      return { deleted: true };
-    } catch (err) {
-      return { deleted: false, message: err.message };
+    const user = await this.userModel.findById(id).exec();
+    if (!user) {
+      throw new NotFoundException('User not found');
     }
+    await this.userModel.findByIdAndDelete(id).exec();
+    return { deleted: true };
   }
 }
